+++ java/azure-perfstress/src/main/java/com/azure/perfstress/SizeOptions.java
@@ -4,5 +4,5 @@ import com.beust.jcommander.Parameter;
 
 public class SizeOptions extends PerfStressOptions {
     @Parameter(names = { "-s", "--size" }, description = "Size of payload (in bytes)")
-    public int Size = 10 * 1024;
+    public long Size = 10 * 1024;
 }
